+++ src/hackingBuddyGPT/usecases/web_api_documentation/__init__.py
@@ -1,0 +1,1 @@
+from .simple_openapi_documentation import SimpleWebAPIDocumentation
