--- src/hackingBuddyGPT/usecases/web_api_testing/testing/__init__.py
@@ -1,1 +1,0 @@
-from .test_handler import GenerationTestHandler
