+++ src/hackingBuddyGPT/utils/web_api/pattern_matcher.py
